Modul 01, draft

+++ chapra_7th/ch03/chapra_example_3_8_float128.py
@@ -1,23 +1,23 @@
 import numpy as np
 
 def calc_quad_root_v1(a, b, c):
-    D = np.float32(b**2) - np.float32(4)*a*c
-    x1 = (-b + np.sqrt(D))/(np.float32(2)*a)
-    x2 = (-b - np.sqrt(D))/(np.float32(2)*a)
+    D = np.float128(b**2) - np.float128(4)*a*c
+    x1 = (-b + np.sqrt(D))/(np.float128(2)*a)
+    x2 = (-b - np.sqrt(D))/(np.float128(2)*a)
     return x1, x2
 
 def calc_quad_root_v2(a, b, c):
-    D = np.float32(b**2) - np.float32(4)*a*c
-    x1 = -np.float32(2)*c/(b + np.sqrt(D))
-    x2 = -np.float32(2)*c/(b - np.sqrt(D))
+    D = np.float128(b**2) - np.float128(4)*a*c
+    x1 = -np.float128(2)*c/(b + np.sqrt(D))
+    x2 = -np.float128(2)*c/(b - np.sqrt(D))
     return x1, x2
 
-a = np.float32(1.0)
-b = np.float32(3000.001)
-c = np.float32(3.0)
+a = np.float128(1.0)
+b = np.float128(3000.001)
+c = np.float128(3.0)
 
-x1_true = np.float32(-0.001)
-x2_true = np.float32(-3000.0)
+x1_true = np.float128(-0.001)
+x2_true = np.float128(-3000.0)
 
 x1, x2 = calc_quad_root_v1(a, b, c)
 print("Using 1st formula: approx roots: ", x1, " ", x2)
